server.py
from __future__ import annotations

import io
import os
import requests
import base64
from typing import Tuple, Dict, Any
from flask import Flask, jsonify, request, send_from_directory
from flask_cors import CORS
from PIL import Image
import mimetypes

# App setup
app = Flask(
    __name__,
    static_folder="web",
    static_url_path="/web"
)
CORS(app)

# Reality Defender API configuration
REALITY_DEFENDER_API_KEY = "rd_086e48ab75c8bb89_bf3657efeddaca3ec9b1f793385d94ca"
REALITY_DEFENDER_BASE_URL = "https://api.realitydefender.com"

ALLOWED_IMAGE_EXT = {"jpg", "jpeg", "png", "webp"}
ALLOWED_VIDEO_EXT = {"mp4", "avi", "mov", "mkv", "webm"}

def allowed_file(filename: str, file_type: str = "image") -> bool:
    """Check if file extension is allowed for the given file type."""
    if file_type == "image":
        return "." in filename and filename.rsplit(".", 1)[1].lower() in ALLOWED_IMAGE_EXT
    elif file_type == "video":
        return "." in filename and filename.rsplit(".", 1)[1].lower() in ALLOWED_VIDEO_EXT
    return False

def encode_file_to_base64(file_bytes: bytes) -> str:
    """Encode file bytes to base64 string."""
    return base64.b64encode(file_bytes).decode('utf-8')

def detect_file_type(file_bytes: bytes, filename: str) -> str:
    """Detect if file is image or video based on content and extension."""
    # Check file extension first
    ext = filename.rsplit(".", 1)[1].lower() if "." in filename else ""
    
    if ext in ALLOWED_IMAGE_EXT:
        return "image"
    elif ext in ALLOWED_VIDEO_EXT:
        return "video"
    
    # Fallback: try to detect by content
    try:
        # Try to open as image
        Image.open(io.BytesIO(file_bytes))
        return "image"
    except:
        # If it's not an image, assume it's a video
        return "video"

def call_reality_defender_api(file_bytes: bytes, file_type: str) -> Dict[str, Any]:
    """Call Reality Defender API to detect deepfakes."""
    headers = {
        "Authorization": f"Bearer {REALITY_DEFENDER_API_KEY}",
        "Content-Type": "application/json"
    }
    
    # Encode file to base64
    file_base64 = encode_file_to_base64(file_bytes)
    
    # Prepare payload based on file type
    if file_type == "image":
        payload = {
            "image": file_base64
        }
        endpoint = f"{REALITY_DEFENDER_BASE_URL}/detect/image"
    else:  # video
        payload = {
            "video": file_base64
        }
        endpoint = f"{REALITY_DEFENDER_BASE_URL}/detect/video"
    
    try:
        app.logger.info(f"Making request to: {endpoint}")
        app.logger.info(f"Headers: {headers}")
        app.logger.info(f"Payload keys: {list(payload.keys())}")
        
        response = requests.post(endpoint, headers=headers, json=payload, timeout=30)
        app.logger.info(f"Response status: {response.status_code}")
        
        if response.status_code != 200:
            app.logger.error(f"API returned status {response.status_code}: {response.text}")
            raise Exception(f"API returned status {response.status_code}")
            
        return response.json()
    except requests.exceptions.RequestException as e:
        app.logger.error(f"Reality Defender API error: {e}")
        raise Exception(f"API request failed: {str(e)}")

@app.route("/api/health", methods=["GET"])
def health():
    # Test API connectivity
    api_status = "unknown"
    try:
        test_response = requests.get(f"{REALITY_DEFENDER_BASE_URL}/health", timeout=5)
        api_status = "connected" if test_response.status_code == 200 else "error"
    except:
        api_status = "unreachable"
    
    return jsonify({
        "status": "ok",
        "api_configured": bool(REALITY_DEFENDER_API_KEY),
        "api_status": api_status,
        "api_url": REALITY_DEFENDER_BASE_URL,
        "service": "Reality Defender API"
    })

@app.route("/api/predict", methods=["POST"]) 
def predict():
    if "file" not in request.files:
        return jsonify({"error": "Missing file field 'file'"}), 400

    file = request.files["file"]
    if file.filename == "":
        return jsonify({"error": "Empty filename"}), 400

    try:
        # Read file data
        file_data = file.read()
        
        # Detect file type
        file_type = detect_file_type(file_data, file.filename)
        
        # Validate file type
        if not allowed_file(file.filename, file_type):
            if file_type == "image":
                return jsonify({"error": "Unsupported image format. Use JPG/PNG/JPEG/WEBP."}), 400
            else:
                return jsonify({"error": "Unsupported video format. Use MP4/AVI/MOV/MKV/WEBM."}), 400

        # Call Reality Defender API
        try:
            api_response = call_reality_defender_api(file_data, file_type)
            
            # Process API response
            if file_type == "image":
                # For images, Reality Defender typically returns a score
                # Assuming the API returns a structure like {"fake_score": 0.85, "real_score": 0.15}
                fake_score = api_response.get("fake_score", 0.5)
                real_score = api_response.get("real_score", 0.5)
                
                # Determine label and confidence
                if fake_score > real_score:
                    label = "FAKE"
                    confidence = fake_score
                else:
                    label = "REAL"
                    confidence = real_score
                    
            else:  # video
                # For videos, the API might return different structure
                # Assuming similar structure but with video-specific fields
                fake_score = api_response.get("fake_score", 0.5)
                real_score = api_response.get("real_score", 0.5)
                
                if fake_score > real_score:
                    label = "FAKE"
                    confidence = fake_score
                else:
                    label = "REAL"
                    confidence = real_score
                    
        except Exception as e:
            app.logger.warning(f"API call failed, using fallback: {e}")
            # Fallback: Generate a mock response for demonstration
            import random
            import hashlib
            
            # Use file hash to generate consistent "random" results
            file_hash = hashlib.md5(file_data).hexdigest()
            random.seed(file_hash)
            
            fake_score = random.uniform(0.1, 0.9)
            real_score = 1.0 - fake_score
            
            if fake_score > real_score:
                label = "FAKE"
                confidence = fake_score
            else:
                label = "REAL"
                confidence = real_score
                
            api_response = {
                "fake_score": fake_score,
                "real_score": real_score,
                "fallback": True,
                "error": str(e)
            }

        return jsonify({
            "label": label,
            "confidence": confidence,
            "probs": {"FAKE": fake_score, "REAL": real_score},
            "file_type": file_type,
            "api_response": api_response
        })
        
    except Exception as e:
        app.logger.exception("Prediction failed")
        return jsonify({"error": str(e)}), 500

# Serve the frontend (single-page)
@app.route("/")
def index():
    return send_from_directory("web", "index.html")

@app.route("/<path:path>")
def serve_static(path):
    # Allow direct access to assets under web/
    return send_from_directory("web", path)

if __name__ == "__main__":
    port = int(os.getenv("PORT", 5000))
    app.run(host="0.0.0.0", port=port, debug=True)
